Average Shopping Bill
You are budgeting your money for the year and must calculate your
average weekly grocery shopping spending over the past five weeks.
Store the past five grocery totals as a list in an array. Create an
expression that will use the items in the array to calculate the
average amount spent on groceries.

Please note: there should only be one given for this problem set, the
array holding the five weekly totals.  You should be able to access the
array using the array access notation discussed earlier in the course.

How to round the x to two decimal places?  Try to find answer.  Tried
Math.round a few ways but didn’t work…