Assignment: Expressions - Personal
Finished - Need one resolution to a step I’m not happy with,
but I have to move on at this point.  For now, it is rigged up to work
using a skip() statement.  Not pretty, but shows green in WebStorm,
throws red error in console, but the script works correctly.  Will come
back to see if I can resolve this after I complete other exercises. In
real world this would definitely need resolved before launch.